Addressing PR comment and moving functions to helper file
